Fix generics

+++ src/test/java/org/marid/typedmap/identity/benchmark/TypedMapPutBenchmark.java
@@ -60,7 +60,7 @@ public class TypedMapPutBenchmark {
     @State(Scope.Thread)
     public static class PutState {
 
-        @Param({"linked", "i16", "chash", "fus"})
+        @Param({"linked", "i32", "chash", "fus"})
         private String type;
 
         Supplier<TypedMutableMap<TestKeyDomain, Integer>> supplier;
